+++ tests/Equinox.EventStore.Integration/Infrastructure.fs
@@ -56,7 +56,7 @@ module SerilogHelpers =
         | _ -> None
     [<RequireQualifiedAccess>]
     type EsAct = Append | AppendConflict | SliceForward | SliceBackward | BatchForward | BatchBackward
-    let (|EsAction|) (evt : Log.Event) =
+    let (|EsAction|) (evt : Log.Metric) =
         match evt with
         | Log.WriteSuccess _ -> EsAct.Append
         | Log.WriteConflict _ -> EsAct.AppendConflict
@@ -64,9 +64,9 @@ module SerilogHelpers =
         | Log.Slice (Direction.Backward,_) -> EsAct.SliceBackward
         | Log.Batch (Direction.Forward,_,_) -> EsAct.BatchForward
         | Log.Batch (Direction.Backward,_,_) -> EsAct.BatchBackward
-    let (|EsEvent|_|) (logEvent : LogEvent) : Log.Event option =
+    let (|EsEvent|_|) (logEvent : LogEvent) : Log.Metric option =
         logEvent.Properties.Values |> Seq.tryPick (function
-            | SerilogScalar (:? Log.Event as e) -> Some e
+            | SerilogScalar (:? Log.Metric as e) -> Some e
             | _ -> None)
 
     let (|HasProp|_|) (name : string) (e : LogEvent) : LogEventPropertyValue option =